requirements.txt
flet>=0.28.0
aiosmtpd>=1.4.4
email-validator>=2.1.0
html2text>=2020.1.16
emoji>=2.8.0
pyinstaller>=6.3.0
pillow>=10.0.0 